reorganized code into loginController, cleaned up loginRoutes
