+++ routes/loginRoutes.js
@@ -1,184 +1,18 @@
 const express = require('express');
-const secrets = require('../secrets');
+const loginRoutes = require('../controllers/loginController');
+
 const querystring = require('querystring');
 const cookieParser = require('cookie-parser');
-const request  = require('request');
-
 
 const router = express.Router();
 
-
-const client_id = secrets.clientID;
-const client_secret = secrets.clientSecret;
-const redirect_uri = "http://localhost:3000/login/callback";
-
-let listOfPlaylists = [];
-
-var generateRandomString = function(length) {
-  var text = '';
-  var possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-
-  for (var i = 0; i < length; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
-};
-
-let stateKey = 'spotify_auth_state';
-
 router.use(cookieParser());
 
-
-
-router.get('/', (req,res) => {
-  res.render('./login/index', { secrets });
-});
-
-router.get('/login', function(req, res) {
-
-  var state = generateRandomString(16);
-  res.cookie(stateKey, state);
-
-  // your application requests authorization
-  var scope = 'user-read-private user-read-email';
-  res.redirect('https://accounts.spotify.com/authorize?' +
-    querystring.stringify({
-      response_type: 'code',
-      client_id: client_id,
-      scope: scope,
-      redirect_uri: redirect_uri,
-      state: state
-    }));
-});
-
-
-router.get('/callback', function(req, res) {
-
-  // your application requests refresh and access tokens
-  // after checking the state parameter
-
-  var code = req.query.code || null;
-  var state = req.query.state || null;
-  var storedState = req.cookies ? req.cookies[stateKey] : null;
-
-  if (state === null || state !== storedState) {
-    res.redirect('/#' +
-      querystring.stringify({
-        error: 'state_mismatch'
-      }));
-  } else {
-    res.clearCookie(stateKey);
-    var authOptions = {
-      url: 'https://accounts.spotify.com/api/token',
-      form: {
-        code: code,
-        redirect_uri: redirect_uri,
-        grant_type: 'authorization_code'
-      },
-      headers: {
-        'Authorization': 'Basic ' + (new Buffer(client_id + ':' + client_secret).toString('base64'))
-      },
-      json: true
-    };
-
-    request.post(authOptions, function(error, response, body) {
-      if (!error && response.statusCode === 200) {
-
-        var access_token = body.access_token,
-            refresh_token = body.refresh_token;
-
-        var options = {
-          url: 'https://api.spotify.com/v1/me',
-          headers: { 'Authorization': 'Bearer ' + access_token },
-          json: true
-        };
-
-        // use the access token to access the Spotify Web API
-        request.get(options, function(error, response, body) {
-          console.log(body);
-          let limit = 50;
-          let offset = 0;
-          let optionsPlaylist = {
-            url: 'https://api.spotify.com/v1/me/playlists?offset=' + offset.toString() + '&limit=' + limit.toString(),
-            headers: { 'Authorization': 'Bearer ' + access_token },
-            json: true
-          }
-
-          request.get(optionsPlaylist, (error, response, body) => {
-
-            console.log(body);
-            body.items.forEach(playlist => {
-              listOfPlaylists.push(playlist);
-            });
-              
-            
-
-
-            if(body.total > limit + offset) {
-              request.get({
-                url: body.next,
-                headers: { 'Authorization': 'Bearer ' + access_token },
-                json: true
-              }, (error, response, body) => {
-
-                console.log(body);
-                console.log(listOfPlaylists);
-              })
-            }
-
-
-          })
-
-
-
-        });
-
-        // we can also pass the token to the browser to make requests from there
-        res.redirect('/login/success/#' +
-          querystring.stringify({
-            access_token: access_token,
-            refresh_token: refresh_token
-          }));
-      } else {
-        res.redirect('/#' +
-          querystring.stringify({
-            error: 'invalid_token'
-          }));
-      }
-    });
-  }
-});
-
-router.get('/refresh_token', function(req, res) {
-
-  // requesting access token from refresh token
-  var refresh_token = req.query.refresh_token;
-  var authOptions = {
-    url: 'https://accounts.spotify.com/api/token',
-    headers: { 'Authorization': 'Basic ' + (new Buffer(client_id + ':' + client_secret).toString('base64')) },
-    form: {
-      grant_type: 'refresh_token',
-      refresh_token: refresh_token
-    },
-    json: true
-  };
-
-  request.post(authOptions, function(error, response, body) {
-    if (!error && response.statusCode === 200) {
-      var access_token = body.access_token;
-      res.send({
-        'access_token': access_token
-      });
-    }
-  });
-});
-
-router.get('/success', (req, res) => {
-  res.render('./login/success', { playlists : listOfPlaylists });
-});
-
-router.get('player/:playerId', (req, res) => {
-  res.render('/login/player')
-});
+router.get('/', loginRoutes.login_index);
+router.get('/login', loginRoutes.login_login);
+router.get('/callback', loginRoutes.login_user);
+router.get('/refresh_token', loginRoutes.login_refreshToken);
+router.get('/playlists', loginRoutes.login_playlists_get);
+router.get('player/:playerId', loginRoutes.login_playerId);
 
 module.exports = router;
